Fix #59. Add unread messages, inline code `sample` and markdown fixes\. This commit add a new feature to the ChatView: unread messages reminder. When you're not at the max scroll (from the bottom), the ChatView will show a little info-bar to indicate that you received a new messages.
Markdown got a new supported tag: inline monospace code blocks.
Just type `var foo = bar;` to have the text displayed in monospace.
This should permit to make more clear conversations for developers ;)

Also, Util.vala was updated to first do emojis then markdown, in order
to not having markdown fucking emojis. Their is still some emojis that
are conflicting with markdown but this should be solved soon as I have
to rewrite all the ChatView code in order to make it efficient.